change primiva type to Object Wrap Class

+++ src/main/java/com/weinyc/sa/core/db/NavigationUtil.java
@@ -30,7 +30,7 @@ public class NavigationUtil {
         bean.setOrder(na.getOrder());
         bean.setId(na.getId().intValue());
         bean.setI18n(na.getI18n());
-        bean.setNavTier(new int[]{(int) na.getTier_1(), (int) na.getTier_2(), (int) na.getTier_3(), (int) na.getTier_4()});
+        bean.setNavTier(new int[]{ na.getTier_1().intValue(), (int) na.getTier_2().intValue(), (int) na.getTier_3().intValue(), (int) na.getTier_4().intValue()});
         return bean;
     }
